refactor: CalculatorDto 이름 변경 및 setter 제거 #1
