--- src/main/java/Calculator/dto/CalculatorDto.java
@@ -1,23 +1,0 @@
-package Calculator.dto;
-
-public class CalculatorDto {
-
-	private String expression;
-	private int result;
-
-	public String getExpression() {
-		return expression;
-	}
-
-	public void setExpression(String expression) {
-		this.expression = expression;
-	}
-
-	public int getResult() {
-		return result;
-	}
-
-	public void setResult(int result) {
-		this.result = result;
-	}
-}
