+++ src/main/java/Calculator/dto/ExpressionDto.java
@@ -1,0 +1,14 @@
+package Calculator.dto;
+
+public class ExpressionDto {
+
+    private String expression;
+
+    public ExpressionDto(String expression) {
+        this.expression = expression;
+    }
+
+    public String getExpression() {
+        return expression;
+    }
+}
